database connected and basic routing setup
