+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "nextjs-app-router-with-firebase-auth",
+  "name": "challery",
   "version": "0.1.0",
   "private": true,
   "scripts": {
