+++ src/app/page.js
@@ -1,44 +1,5 @@
-/* eslint-disable react-hooks/exhaustive-deps */
+import React from 'react';
 
-'use client';
-
-import React, { useEffect, useState } from 'react';
-import Link from 'next/link';
-import { Button } from 'react-bootstrap';
-import { getBooks } from '../api/bookData';
-import { useAuth } from '../utils/context/authContext';
-import BookCard from '../components/BookCard';
-
-function Home() {
-  // TODO: Set a state for books
-  const [books, setBooks] = useState([]);
-
-  // TODO: Get user ID using useAuth Hook
-  const { user } = useAuth();
-
-  // TODO: create a function that makes the API call to get all the books
-  const getAllTheBooks = () => {
-    getBooks(user.uid).then(setBooks);
-  };
-
-  // TODO: make the call to the API to get all the books on component render
-  useEffect(() => {
-    getAllTheBooks();
-  }, []);
-
-  return (
-    <div className="text-center my-4">
-      <Link href="/book/new" passHref>
-        <Button>Add A Book</Button>
-      </Link>
-      <div className="d-flex flex-wrap">
-        {/* TODO: map over books here using BookCard component */}
-        {books.map((book) => (
-          <BookCard key={book.firebaseKey} bookObj={book} onUpdate={getAllTheBooks} />
-        ))}
-      </div>
-    </div>
-  );
+export default function Landing() {
+  return <div>Landing Page</div>;
 }
-
-export default Home;
